+++ src/party_menu.c
@@ -1908,20 +1908,17 @@ static bool8 CanLearnTutorMove(u16 species, u8 tutor)
     switch (tutor)
     {
     case TUTOR_MOVE_FRENZY_PLANT:
-        if (species == SPECIES_VENUSAUR)
-            return TRUE;
-        else
-            return FALSE;
+        return (species == SPECIES_VENUSAUR
+             || species == SPECIES_MEGANIUM
+             || species == SPECIES_SCEPTILE)
     case TUTOR_MOVE_BLAST_BURN:
-        if (species == SPECIES_CHARIZARD)
-            return TRUE;
-        else
-            return FALSE;
+        return (species == SPECIES_CHARIZARD
+             || species == SPECIES_TYPHLOSION
+             || species == SPECIES_BLAZIKEN)
     case TUTOR_MOVE_HYDRO_CANNON:
-        if (species == SPECIES_BLASTOISE)
-            return TRUE;
-        else
-            return FALSE;
+        return (species == SPECIES_BLASTOISE
+             || species == SPECIES_FERALIGATR
+             || species == SPECIES_SWAMPERT)
     default:
         if (sTutorLearnsets[species] & (1 << tutor))
             return TRUE;